+++ thriftrw/wire/__init__.py
@@ -22,8 +22,7 @@
 Types
 -----
 
-.. autodata:: thriftrw.wire.TType
-    :annotation:
+.. automodule:: thriftrw.wire.ttype
 
 Value
 -----
@@ -65,8 +64,6 @@ Value Types
 """
 from __future__ import absolute_import, unicode_literals, print_function
 
-from .ttype import TType
-
 from .value import (
     Value,
     BoolValue,
@@ -85,7 +82,7 @@ from .value import (
 )
 
 __all__ = [
-    'TType',
+    'ttype',
 
     # Value
     'Value',